Re-coded how task interfaces with JSMin.java, now passes input and output files through with InputFileStream and OutputFileStream. This means we can use the original JSMin.java code without it being altered.
Added destdir param, this allows users o specify a destination driectory for the files to be outputted to. Defaults to directory of the source file.

Added suffix param, this allows the user to decide whether they want to add the .min.js file exstension or just use the original file name. Default is no .min.js suffix.